docs: establish naming convention for examples and tutorials
Apply consistent naming across examples/ and tutorials/:

- Delete stale _main duplicates in unicycle/ and van_der_pol/
- Rename _control.py files to descriptive method names (e.g., ekf.py,
  vanilla_cbf.py, mppi_cbf.py) across all example directories
- Reorganize differential_drive/ into obstacle_avoidance/ and
  human_aware_navigation/ task subdirectories
- Reorganize pedestrian/ into navigate_among_pedestrians/ with
  scenario-descriptive names (crowded.py, head_on.py, etc.)
- Rename tutorials: drop simulate_ prefix, delete stale duplicates
- Rename adaptive_cvar_cbf demo to match convention
- Update tests, README, .gitignore to reflect new paths
- Update renaming_plan.md with formal convention and completed status
