+++ .gitignore
@@ -185,7 +185,7 @@ gemini.md
 examples/unicycle/reach_goal/results/bh_mppi_cbf_control.eps
 examples/unicycle/reach_goal/results/bh_mppi_cbf_control.mp4
 examples/unicycle/reach_goal/results/bh_mppi_cbf_control.eps
-examples/differential_drive/results/augmented_dynamic_animation.mp4
+examples/differential_drive/obstacle_avoidance/results/augmented_dynamic_animation.mp4
 tutorials/tutorials/van_der_pol_oscillator
 tutorials/tutorials/mppi_cbf_stl
 AGENTS.md
@@ -194,16 +194,15 @@ examples/single_integrator/reach_goal/ukf_state_estimation/test.py
 debug_manager.py
 examples/adaptive_cvar_cbf/results/cbfkit_demo.mp4
 examples/adaptive_cvar_cbf/results/simulation.mp4
-examples/differential_drive/results/dynamic_obstacle_animation.mp4
-examples/differential_drive/results/human_aware_animation.mp4
-examples/differential_drive/results/multi_scenario_animation.mp4
-examples/differential_drive/results/single_robot_cbf_animation.mp4
-examples/pedestrian/results/crowded_demo.mp4
-examples/pedestrian/results/crowded_naturalistic_demo.mp4
-examples/pedestrian/results/head_on_demo.mp4
-examples/pedestrian/results/overtaking_demo.mp4
-examples/pedestrian/results/pedestrian_manager_demo.mp4
-examples/pedestrian/results/star_burst_demo.mp4
+examples/differential_drive/obstacle_avoidance/results/dynamic_obstacle_animation.mp4
+examples/differential_drive/human_aware_navigation/results/human_aware_animation.mp4
+examples/differential_drive/human_aware_navigation/results/multi_scenario_animation.mp4
+examples/differential_drive/obstacle_avoidance/results/single_robot_cbf_animation.mp4
+examples/pedestrian/navigate_among_pedestrians/results/crowded.mp4
+examples/pedestrian/navigate_among_pedestrians/results/head_on.mp4
+examples/pedestrian/navigate_among_pedestrians/results/overtaking.mp4
+examples/pedestrian/navigate_among_pedestrians/results/crossing.mp4
+examples/pedestrian/navigate_among_pedestrians/results/star_burst.mp4
 examples/single_integrator/reach_goal/ekf_state_estimation/test.py
 examples/single_integrator/reach_goal/ekf_state_estimation/results/ra_fxt_clf_n100_pg55_controls.eps
 examples/single_integrator/reach_goal/ekf_state_estimation/results/ra_fxt_clf_n100_pg55_lyapunov_vals.eps
@@ -211,7 +210,7 @@ examples/single_integrator/reach_goal/ekf_state_estimation/results/ra_fxt_clf_n1
 examples/van_der_pol/regulation/perfect_state_measurements/test.py
 tutorials/test.py
 REFACTORING_REPORT.md
-examples/differential_drive/results/navigation_animation.mp4
+examples/differential_drive/obstacle_avoidance/results/navigation_animation.mp4
 examples/unicycle/reach_goal/results/bh_mppi_cbf_control.eps
 examples/unicycle/reach_goal/results/bh_mppi_cbf_control.mp4
 examples/unicycle/reach_goal/results/bh_mppi_cbf_control.eps
