+++ examples/differential_drive/obstacle_avoidance/augmented_dynamic_obstacle_cbf.py
@@ -235,7 +235,7 @@ def run_simulation():
         nominal_controller=nom_controller,
         sensor=sensor,
         estimator=estimator,
-        filepath="examples/differential_drive/results/augmented_results",
+        filepath="examples/differential_drive/obstacle_avoidance/results/augmented_results",
         verbose=True,
     )
 
@@ -300,7 +300,7 @@ def create_visualization(states, goal_state, d_min):
 
     anim = animation.FuncAnimation(fig, animate, frames=len(states), interval=50, blit=True)
 
-    save_path = "examples/differential_drive/results/augmented_dynamic_animation.mp4"
+    save_path = "examples/differential_drive/obstacle_avoidance/results/augmented_dynamic_animation.mp4"
     try:
         anim.save(save_path, writer="ffmpeg", fps=20)
         print(f"Saved to {save_path}")
