+++ examples/pedestrian/navigate_among_pedestrians/crossing.py
@@ -1,19 +1,21 @@
 """
-Pedestrian Head-On Interaction Demo.
+Pedestrian Manager Demo.
 
-Demonstrates the robot navigating a head-on encounter with a pedestrian.
+Demonstrates the use of `cbfkit.systems.pedestrian.CrowdManager` to easily setup
+a robot navigation scenario with multiple pedestrians using Social Force behavior.
+
+Scenario: 'Crossing'
+- Robot starts at (0, 5) heading Right.
+- 2 Pedestrians cross vertically.
 """
 
 import os
 import sys
-from pathlib import Path
 
 import jax.numpy as jnp
-import numpy as np
 
 # Add project root
-PROJECT_ROOT = Path(__file__).resolve().parent.parent.parent
-sys.path.append(str(PROJECT_ROOT))
+sys.path.append(os.getcwd())
 
 from jax import jit
 
@@ -23,6 +25,8 @@ from cbfkit.controllers.mppi.mppi_generator import mppi_generator
 from cbfkit.estimators import naive as estimator
 from cbfkit.integration import runge_kutta_4 as integrator
 from cbfkit.sensors import perfect as sensor
+
+# --- New Imports ---
 from cbfkit.systems.pedestrian import CrowdManager
 from cbfkit.systems.pedestrian.behaviors import social_force_policy
 from cbfkit.utils.user_types import ControllerData, PlannerData
@@ -30,20 +34,26 @@ from cbfkit.utils.visualization import visualize_crowd
 
 
 def run_demo():
-    print("Initializing Head-On Demo...")
+    print("Initializing CrowdManager Demo...")
 
+    # 1. Setup Manager
     manager = CrowdManager()
 
-    # Pedestrian moving Left towards robot
+    # 2. Add Pedestrians (Social Force)
+    # Pedestrian 1: Moving Up
     manager.add_pedestrian(
-        init_state=jnp.array([10.0, 5.0, -1.0, 0.0]),
-        behavior=social_force_policy(goal=jnp.array([0.0, 5.0]), desired_speed=1.0),
-        id="ped_head_on",
+        init_state=[5.0, 0.0, 0.0, 0],
+        behavior=social_force_policy(goal=jnp.array([0.0, 0.0]), desired_speed=1.0),
+    )
+    # Pedestrian 2: Moving Down
+    manager.add_pedestrian(
+        init_state=[10.0, 10.0, 0.0, 0],
+        behavior=social_force_policy(goal=jnp.array([0.0, 0.0]), desired_speed=1.0),
     )
 
     num_peds = len(manager.pedestrians)
 
-    # Robot moving Right towards pedestrian
+    # 3. Robot Setup
     robot_dyn = unicycle.plant(l=1.0)
     robot_dyn.a_max = 5.0
     robot_dyn.omega_max = 5.0
@@ -52,23 +62,25 @@ def run_demo():
     x0_robot = jnp.array([0.0, 5.0, 0.0, 0.0])
     goal_robot = jnp.array([10.0, 5.0, 0.0, 0.0])
 
-    # Augmented System
+    # 4. Get Augmented System
     aug_dynamics = manager.get_augmented_dynamics(robot_dyn)
     closed_loop_dynamics = manager.get_closed_loop_dynamics(robot_dyn)
     z0 = manager.get_initial_state(x0_robot)
 
+    # 5. Configure MPPI (Nominal Planner)
     # MPPI Costs
     @jit
     def stage_cost(z, u):
         p_r = z[:2]
         dist_goal = jnp.linalg.norm(p_r - goal_robot[:2])
 
+        # Simple Obstacle Repulsion (based on augmented state z)
         c_obs = 0.0
         for i in range(num_peds):
             idx = 4 + i * 4
             p_h = z[idx : idx + 2]
             dist_h = jnp.linalg.norm(p_r - p_h)
-            c_obs += 150.0 * jnp.exp(-2.5 * (dist_h - 1.0))  # Stronger repulsion for head-on
+            c_obs += 100.0 * jnp.exp(-2.0 * (dist_h - 1.0))
 
         return 0.01 * jnp.dot(u, u) + 2.0 * dist_goal + c_obs
 
@@ -78,7 +90,7 @@ def run_demo():
         return 10.0 * jnp.linalg.norm(p_r - goal_robot[:2]) ** 2
 
     mppi_params = {
-        "prediction_horizon": 25,
+        "prediction_horizon": 20,
         "num_samples": 500,
         "time_step": 0.1,
         "use_GPU": True,
@@ -89,6 +101,7 @@ def run_demo():
         "cost_perturbation": 0.1,
     }
 
+    # Use closed_loop_dynamics for planning
     planner = mppi_generator()(
         control_limits=jnp.array([5.0, 5.0]),
         dynamics_func=closed_loop_dynamics,
@@ -97,11 +110,22 @@ def run_demo():
         mppi_args=mppi_params,
     )
 
-    init_planner_data = PlannerData(u_traj=jnp.zeros((25, 2)))
+    init_planner_data = PlannerData(u_traj=jnp.zeros((20, 2)))
+
+    # Wrap planner data in ControllerData because we use MPPI as controller
+    # Use 'inner_controller_data' key for CrowdManager to unpack
     init_controller_data = ControllerData(sub_data={"inner_controller_data": init_planner_data})
 
+    # 6. Wrap Controller
     combined_controller = manager.get_nominal_controller(planner, use_augmented_state=True)
 
+    # 7. CBF Safety Filter (Optional but Recommended)
+    # For simplicity in this demo, we might skip CBF or add it.
+    # If we add it, we must wrap the CBF controller which wraps the Nominal Controller.
+    # Let's stick to MPPI only for this demo to verify the CrowdManager logic first.
+    # If MPPI avoids, the dynamics/manager integration is correct.
+
+    # 8. Run Simulation
     dt = 0.1
     if os.getenv("CBFKIT_TEST_MODE"):
         tf = 1.0
@@ -115,15 +139,30 @@ def run_demo():
         num_steps=int(tf / dt),
         dynamics=aug_dynamics,
         integrator=integrator,
+        # Note: We pass 'combined_controller' as the primary 'controller' here if we don't use CBF.
+        # If we used CBF, we'd pass CBF as controller and combined_controller as nominal.
+        # But `sim.execute` logic calls `controller`.
         controller=combined_controller,
         sensor=sensor,
         estimator=estimator,
-        controller_data=init_controller_data,
+        controller_data=init_controller_data,  # Pass wrapped data
         verbose=True,
+        use_jit=True,
     )
 
-    results_dir = Path(__file__).parent / "results"
-    results_dir.mkdir(parents=True, exist_ok=True)
+    # 9. Visualize
+    # Unpack p_values from controller_data if needed?
+    # sim.execute returns p_keys, p_values if planner is used.
+    # Here planner is NONE. So p_keys/values will be empty.
+    # The planner data is hidden inside c_values (ControllerData).
+    # c_keys will contain 'sub_data' maybe?
+    # Actually, sim.execute logs controller_data fields.
+
+    # We need to extract MPPI trajectories for visualization from c_values.
+    # This might be tricky if c_values flattens sub_data.
+    # Let's see what we get.
+
+    os.makedirs("examples/pedestrian/results", exist_ok=True)
     if not os.getenv("CBFKIT_TEST_MODE"):
         visualize_crowd(
             states=x,
@@ -131,9 +170,9 @@ def run_demo():
             robot_goal=goal_robot,
             d_safe=1.0,
             dt=dt,
-            p_values=p_values,
-            p_keys=p_keys,
-            save_path=str(results_dir / "head_on_demo.mp4"),
+            # p_values=p_values, # Likely empty
+            # p_keys=p_keys,
+            save_path="examples/pedestrian/navigate_among_pedestrians/results/crossing.mp4",
         )
 
     print("Demo Complete!")
