+++ examples/unicycle/reach_goal/perfect_sensing.py
@@ -87,5 +87,5 @@ if plot:
         dt=dt,
         title="System Behavior",
         save_animation=save,
-        animation_filename="examples/unicycle/start_to_goal/results/perfect_measurements.gif",
+        animation_filename="examples/unicycle/reach_goal/results/perfect_sensing.gif",
     )
