+++ examples/unicycle/reach_goal/risk_aware_cbf.py
@@ -74,5 +74,5 @@ if animate:
         dt=dt,
         title="System Behavior",
         save_animation=save,
-        animation_filename="examples/unicycle/reach_goal/results/start_to_goal1.gif",
+        animation_filename="examples/unicycle/reach_goal/results/risk_aware_cbf.gif",
     )
